Trying to fix weird problem with patch
I tried to use the patch file from "bebosudo/olive-editor" repo and it
wasn't applied, but if I change the name of the patch on my repo and try
to apply it works. I don't know why, but even the two files being
identical, if I run "diff" it says that the two files are different, and
that may be why the patch fails.